internal/*: drop merging authorized_keys.d into authorized_keys
This makes it a default that the SSH key fragments in
.ssh/authorized_keys.d are not synced to .ssh/authorized_keys.
Ignition will write out key fragments only in .ssh/authorized_keys.d.

The flag useAuthorizedKeysFile is added to the distro package. If
this flag is set to "true", then the previous behavior of syncing
the key fragments to authorized_keys is maintained. The reason for
adding the flag is so that distros not supporting authorized_keys.d
can still use authorized_keys.

Fixes: #716